refactor preference strings into separate module
