+++ ts/speech_generator/speech_generator_util.ts
@@ -32,6 +32,7 @@ import { SemanticTree } from '../semantic_tree/semantic_tree.js';
 import * as WalkerUtil from '../walker/walker_util.js';
 import * as EngineConst from '../common/engine_const.js';
 import { ClearspeakPreferences } from '../speech_rules/clearspeak_preferences.js';
+import { addPreference, findPreference } from '../speech_rules/clearspeak_preference_string.js';
 // import { RebuildStree } from '../walker/rebuild_stree.js';
 
 type OptionsList = { [key: string]: string };
@@ -485,12 +486,13 @@ export function nextRules(
  * @returns The new style name.
  */
 export function nextStyle(node: SemanticNode, options: OptionsList) {
-  const { modality: modality, domain: domain, style: style } = options;
+  const { modality: modality, domain: domain, style: style, locale: locale } = options;
   // Rule cycling only makes sense for speech modality.
   if (modality !== 'speech') {
     return style;
   }
 
+  console.log(`${locale} ${domain} ${style}`);
   if (domain === 'mathspeak') {
     const styles = ['default', 'brief', 'sbrief'];
     const index = styles.indexOf(style);
@@ -501,35 +503,40 @@ export function nextStyle(node: SemanticNode, options: OptionsList) {
   }
   if (domain === 'clearspeak') {
     const prefs = ClearspeakPreferences.getLocalePreferences();
-    const loc = prefs['en'];
-    // TODO: use correct locale.
+    const loc = prefs[locale];
     if (!loc) {
       return 'default';
     }
     // TODO: return the previous one?
     const smart = ClearspeakPreferences.relevantPreferences(node);
-    const current = ClearspeakPreferences.findPreference(style, smart);
+    const current = findPreference(style, smart);
     const options = loc[smart].map(function (x) {
       return x.split('_')[1];
     });
     const index = options.indexOf(current);
+    console.log(current);
     if (index === -1) {
+      console.log(31);
       return style;
     }
     const next = index >= options.length - 1 ? options[0] : options[index + 1];
-    const result = ClearspeakPreferences.addPreference(style, smart, next);
+    const result = addPreference(style, smart, next);
+    console.log(32);
+    console.log(result);
     return result;
   }
   return style;
 }
 
-export function toStyles(option: string): OptionsList {
+export function toStyles(options: OptionsList): OptionsList {
+  const {domain, style, domain2style} = options;
   const styles: OptionsList = {};
-  if (!option) {
+  if (!domain2style) {
     Object.assign(styles, EngineConst.DOMAIN_TO_STYLES);
+    styles[domain] = style;
     return styles;
   }
-  const split = option.split(',');
+  const split = domain2style.split(',');
   for (const pair of split) {
     const [first, second] = pair.split(/:(.*)/);
     styles[first] = second ? second :
